Fix keyboard repeat rate resetting when re-plugging in keyboard:
Apparently Linux treats the keyboard as a "new keyboard", and does
not assign the same settings to it. There is this cool program called
inputplug which responds to different X events, specifically related
to devices being plugged and unplugged. This script will just respond
to all plugin-related events, which is fine for now, but could be
optimized.